% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutorial.R
\name{crimemapping_tutorial}
\alias{crimemapping_tutorial}
\title{Run tutorial}
\usage{
crimemapping_tutorial(name = NULL)
}
\arguments{
\item{name}{Name of the tutorial to be loaded}
}
\description{
This function runs a crime mapping tutorial.
}
