% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tutorial.R
\name{tutorial}
\alias{tutorial}
\title{Run tutorial}
\usage{
tutorial(name = NULL)
}
\arguments{
\item{name}{Name of the tutorial to be loaded}
}
\description{
This function runs a crime mapping tutorial.
}
