% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update tutorials}
\usage{
update(...)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[remotes]{install_github}}.}
}
\description{
This function updates the crime mapping tutorials.
}
