% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_code.R
\name{check_code}
\alias{check_code}
\title{Check code file before submission}
\usage{
check_code(file, reprex = TRUE, style = TRUE)
}
\arguments{
\item{file}{A file path for the .R file to check}

\item{reprex}{Should the code be checked for errors?}

\item{style}{Should the code be checked for style issues?}
}
\value{
NULL, invisibly
}
\description{
Check your code to identify various common issues with code content, code
style, etc.
}
